Edited the existing content which referred to the global alert to be for the in-page alert a22-docs.stories.mdx
Content:
In-page alerts are designed to capture the attention of the user in a deliberately way on a specific page.

Usage
Use in-page alerts to attract attention to an important time sensitive messages, to encourage an action, or to inform the user about the status of an interaction on a single page.

For example, you might use an in-page alert:

When a user submits a form, to tell them that their submission has been successfully received and they can exit the page now.
To inform users of a temporary change to a service that is described on the page.
To inform the user of a system error.
Unlike the global alert, in-page alerts cannot be dismissed or closed by the user.

Do
Use clear, concise language to minimise cognitive load.
Display the alerts in context and at an appropriate time.
Use the appropriate alert colour for the type and urgency of the alert.
Don't
Overuse alerts, as this will erode their effectiveness.
Use for information that is already communicated in a global alert.
In-page alert types
The in-page alert has 4 colour styles which can be used to show different kinds of message.

The styles are:

Information alert (blue)
Designed to be less intrusive or interruptive than other levels of alert, to draw attention to an important message that is not urgent or related to the status of an interaction.

Use for information the user should know, but is not critical.
Use for tips or information which the user can benefit from.
Don't use in response to a user action.
Success alert (green)
Use to inform the user that an action was performed successfully (e.g. submitting a form or registering an account).

Warning alert (orange)
Use to warn the user of a possible negative outcome (e.g. password expiry).
Provide sufficient information to avoid the problem.
Use for an action that is out of the ordinary or might not be desired.
Critical alert (red)
Use where a system event has failed.
Use when the user has made an error.
Use where the user is blocked until the issue is resolved, or the issue needs resolving immediately.
Don't use for validation of an action.
When using the critical alert to inform the user that they need to resolve a problem, make sure the user is clearly shown what the issue is and how to resolve it.